fix: 초기 메인 페이지 수정
