+++ app/app/(tabs)/_layout.tsx
@@ -1,35 +1,30 @@
-import { HapticTab } from '@/components/HapticTab';
 import { IconSymbol } from '@/components/ui/IconSymbol';
-import TabBarBackground from '@/components/ui/TabBarBackground';
-import { Colors } from '@/constants/Colors';
 import { useColorScheme } from '@/hooks/useColorScheme';
 import { Ionicons } from '@expo/vector-icons';
-import { Tabs, usePathname, useSegments } from 'expo-router';
+import { Tabs, useSegments } from 'expo-router';
 import React from 'react';
-import { Platform } from 'react-native';
 
-const hideTabBarScreens = ['(single-running)', '(group-running)'];
+const hideTabBarScreens = ['(single-running)'];
 
 export default function TabLayout() {
-  const colorScheme = useColorScheme();
   const segments = useSegments(); // usePathname 대신 useSegments 사용
 
   // segments에서 현재 탭 정보 추출
   const currentTab = segments[1] || ''; // (tabs) 다음 세그먼트가 현재 탭
-  const fullPath = `/${segments.join('/')}`;
 
   const isHideTabBar = hideTabBarScreens.some(screen =>
     currentTab.includes(screen.replace(/[()]/g, ''))
   );
+  //시작시 경로 home으로 설정
   return (
     <Tabs
       backBehavior="history"
+      initialRouteName="(home)"
       screenOptions={{
         headerShown: false,
         tabBarStyle: {
           display: isHideTabBar ? 'none' : 'flex'
         }
-        // tabBarButton: (props) => <AnimatedTabBarButton {...props} />,
       }}
     >
       <Tabs.Screen
@@ -61,12 +56,6 @@ export default function TabLayout() {
       />
       <Tabs.Screen
         name="(single-running)"
-        options={{
-          href: null // 탭바에는 표시 안 함
-        }}
-      />
-      <Tabs.Screen
-        name="(group-running)"
         options={{
           href: null // 탭바에는 표시 안 함
         }}
